update func_scanner & buf_vuln_scan
1.  func_scanner增加了宽字符类型识别。
2. buf_vuln_scan增加了宽字符字符串常量识别。
3. buf_vuln_scan实现了"strcpy", "wcscpy", "strncpy", "wcsncpy", "memcpy", "memset"函数的漏洞检测，通过单测。

+++ kernel/func_scanner.h
@@ -31,7 +31,7 @@ struct ValueInfo {
     // for pointer or array
     bool is_pointer_;
     bool is_array_;
-    int len_;
+    size_t len_;
 
     ValueInfo(string name, size_t start, size_t end, int deep, bool unsign, string type, int size,
         bool is_pointer, bool is_array, int len) :
